Fixed old web scraper not working
